consolidation mock's files
