+++ src/constants/mocks/courseCompletionChartMock.ts
