+++ src/constants/mocks/courseCompletionRateChartMock.ts
